Began refactoring component layouts for better project organization. Moved input and image form elements to different section of page
