+++ package.json
@@ -10,6 +10,7 @@
     "material-ui-icons": "^1.0.0-beta.35",
     "prop-types": "^15.6.0",
     "react": "^16.2.0",
+    "react-custom-scrollbar": "^1.0.0",
     "react-custom-scrollbars": "^4.2.1",
     "react-dom": "^16.2.0",
     "react-dropzone": "^4.2.8",
