+++ src/components/Alert/Styled.jsx
@@ -1,5 +1,5 @@
 import styled from 'styled-components';
-import constants from '../../constants';
+import constants from '../constants';
 
 export const MessageContainer = styled.div`
   width: 100%;
