+++ src/components/Announce/index.js
@@ -1,0 +1,3 @@
+import Announce from './Announce';
+
+export default Announce;
