+++ src/components/Breadcrumbs/Styled.jsx
@@ -1,0 +1,15 @@
+import styled from 'styled-components';
+
+export const Container = styled.ul`
+  display: flex;
+  color: white;
+  list-style-type: none;
+  padding: 0;
+  margin: 30px 0 0 0;
+
+  li {
+    display: flex;
+    line-height: 24px;
+    margin-right: 10px;
+  }
+`;
