+++ src/components/Breadcrumbs/index.js
@@ -1,0 +1,3 @@
+import Breadcrumbs from './Breadcrumbs';
+
+export default Breadcrumbs;
