+++ src/components/ColorSelect/Styled.jsx
@@ -8,6 +8,7 @@ export const Container = styled.div`
   border: 1px solid rgba(250, 250, 250, 0.1);
   box-shadow: 0 3px 12px rgba(0, 0, 0, 0.3);
   position: absolute;
+  z-index: 1000;
 `;
 
 export const ColorBox = styled.div`
