+++ src/components/Design/index.js
@@ -1,0 +1,3 @@
+import Design from './Design';
+
+export default Design;
