+++ src/components/Products/Products.jsx
@@ -1,13 +1,32 @@
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
+import { CircularProgress } from 'material-ui/Progress';
 import Product from '../Product';
 import { Wrapper, Container, Heading, FlexContainer, Hero } from './Styled';
 import { getProducts } from '../../actions/product';
+import { logout } from '../../actions/auth';
 
-class ProductsPage extends Component {
+class Products extends Component {
+  constructor() {
+    super();
+
+    this.state = {
+      loading: true,
+    };
+  }
+
   componentWillMount() {
-    this.props.getProducts();
+    this.props
+      .getProducts()
+      .then(() => {
+        setTimeout(() => {
+          this.setState({ loading: false });
+        }, 5000);
+      })
+      .catch(() => {
+        this.props.logout();
+      });
   }
 
   render() {
@@ -26,17 +45,26 @@ class ProductsPage extends Component {
     return (
       <Wrapper alertMessage={showAlertMessage}>
         <Hero img="https://images.unsplash.com/photo-1496442226666-8d4d0e62e6e9?ixlib=rb-0.3.5&s=ce0f350894216f6c6f9218865b3db36d&auto=format&fit=crop&w=1350&q=80" />
-        <Container>
-          <Heading>Products</Heading>
-          <FlexContainer>{productElements}</FlexContainer>
-        </Container>
+        {this.state.loading ? (
+          <Container>
+            <CircularProgress
+              color="secondary"
+              style={{ position: 'relative', left: '50%', transform: 'translateX(-50%)', marginTop: '100px' }}
+            />
+          </Container>
+        ) : (
+          <Container>
+            <Heading>Products</Heading>
+            <FlexContainer>{productElements}</FlexContainer>
+          </Container>
+        )}
       </Wrapper>
     );
   }
 }
 
 const { bool, func, arrayOf, shape, string } = PropTypes;
-ProductsPage.propTypes = {
+Products.propTypes = {
   getProducts: func.isRequired,
   products: arrayOf(
     shape({
@@ -46,6 +74,7 @@ ProductsPage.propTypes = {
     })
   ).isRequired,
   showAlertMessage: bool.isRequired,
+  logout: func.isRequired,
 };
 
 function mapStateToProps(state) {
@@ -56,4 +85,4 @@ function mapStateToProps(state) {
   };
 }
 
-export default connect(mapStateToProps, { getProducts })(ProductsPage);
+export default connect(mapStateToProps, { getProducts, logout })(Products);
