+++ src/components/Products/index.js
@@ -1,0 +1,3 @@
+import Products from './Products';
+
+export default Products;
