+++ src/components/inputs/InputGroup/Styled.jsx
@@ -6,7 +6,7 @@ export const Container = styled.div`
   font-family: ${constants.fontFamily};
   font-size: 12px;
   margin-top: 15px;
-  padding: 0 16px;
+  padding: 0;
 
   & > input {
     margin-top: 10px;
