--- src/components/messages/AlertMessage/index.js
@@ -1,3 +1,0 @@
-import AlertMessage from './AlertMessage';
-
-export default AlertMessage;
